updated models/utils.js by moving some functions that were either duplicates or not part of the exported model in models/*.js to utils.js
