+++ backend/models/booking.js
@@ -3,14 +3,9 @@ const Booking = require("../orm/model-router.js")("booking");
 const Bike = require("../orm/model-router.js")("bike");
 const User = require("../orm/model-router.js")("user");
 const Invoice = require("../orm/model-router.js")("invoice");
-
+const { getRandomInt } = require("./utils.js")
 const { Op } = require("sequelize");
 
-function getRandomInt(min, max) {
-    min = Math.ceil(min);
-    max = Math.floor(max);
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-}
 
 const booking = {
     /**
